Minor revisions - wording changes for clarity throughout, biggest change is lack of website due to RSC rebuild, removed references to this data but need to decide whether to stand up and maintain the Autotune data and parameter website if we reinsert this to strengthen the publication.
Need to insert more info in resubmission to explicitly address reviewers comments regarding methodology:
The paper is well organized, but the introduction and background somehow should be shorten.I agree with the authors that the computational issue is an important in building model tuning. However, in its current version, the papers lacks of scientific bases, i.e. detailed explanation of the adopted methodologies. For example, the authors show that the performance of the tuning using both schedules in parallel was competitive with the full schedule approach, but we do not know why? I therefore suggested the authors to provide more details on those methods in the chapter of 'methodology'.